requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pillow==9.5.0
numpy==1.25.2
jinja2==3.1.2
python-multipart
scikit-learn==1.2.2
keras==2.15.0
tensorflow==2.15.0